Fix nswag build and package generation
This also removes support for VS2017 since it was getting hard to keep
both working. It changes all the projects to a single *.sln structure, and
it moves all of the build functionality over to msbuild in the form of the
EmsApi.targets file.
