--- csharp/EmsApi.Client/EmsApi.Client.Vs2015.csproj
@@ -1,92 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <MinimumVisualStudioVersion>14.0</MinimumVisualStudioVersion>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{60E9EF36-B92A-4486-A1C1-BD90C44188C4}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>EmsApi.Client</RootNamespace>
-    <AssemblyName>EmsApi.Client</AssemblyName>
-    <DefaultLanguage>en-US</DefaultLanguage>
-    <FileAlignment>512</FileAlignment>
-    <ProjectTypeGuids>{786C830F-07A1-408B-BD7F-6EE04809D6DB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
-    <TargetFrameworkProfile>
-    </TargetFrameworkProfile>
-    <TargetFrameworkVersion>v5.0</TargetFrameworkVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <DocumentationFile>bin\Debug\EmsApi.Client.xml</DocumentationFile>
-    <NoWarn>CS1591</NoWarn>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <DocumentationFile>bin\Release\EmsApi.Client.xml</DocumentationFile>
-    <NoWarn>CS1591</NoWarn>
-  </PropertyGroup>
-  <ItemGroup>
-    <None Include="project.json" />
-    <!-- A reference to the entire .NET Framework is automatically included -->
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="V2\Access\AssetsAccess.cs" />
-    <Compile Include="V2\Access\EmsApiRouteAccess.cs" />
-    <Compile Include="V2\Access\EmsSystemsAccess.cs" />
-    <Compile Include="V2\Access\SwaggerAccess.cs" />
-    <Compile Include="V2\IEmsApi.cs" />
-    <Compile Include="V2\Model\Aircraft.cs" />
-    <Compile Include="V2\Model\Airport.cs" />
-    <Compile Include="V2\Model\CommonModels.cs" />
-    <Compile Include="V2\Model\EmsSystem.cs" />
-    <Compile Include="V2\Model\EmsSystemInfo.cs" />
-    <Compile Include="V2\Model\Fleet.cs" />
-    <Compile Include="V2\Model\FlightPhase.cs" />
-    <Compile Include="V2\Service\ApiEventArgs.cs" />
-    <Compile Include="V2\Service\Authentication.cs" />
-    <Compile Include="V2\Service\EmsApiEndpoints.cs" />
-    <Compile Include="V2\Service\EmsApiService.cs" />
-    <Compile Include="V2\Service\EmsApiServiceConfiguration.cs" />
-    <Compile Include="V2\Service\EmsApiServiceException.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Properties\" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\EmsApi.Dto\EmsApi.Dto.csproj">
-      <Project>{86f1df15-9bc1-4828-80db-878db1db037d}</Project>
-      <Name>EmsApi.Dto</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
-  <PropertyGroup>
-    <PostBuildEvent>copy "$(TargetPath)" ..\
-copy "$(TargetDir)EmsApi.Client.xml" ..\
-mkdir "$(ProjectDir)..\bin\"
-copy "$(TargetPath)" "$(ProjectDir)..\bin\"
-</PostBuildEvent>
-  </PropertyGroup>
-  <PropertyGroup>
-    <PreBuildEvent>echo $(MSBuildToolsPath)</PreBuildEvent>
-  </PropertyGroup>
-  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
-       Other similar extension points exist, see Microsoft.Common.targets.
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-  -->
-</Project>
