--- csharp/EmsApi.Client/EmsApi.Client.Vs2015.sln
@@ -1,50 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Client.Vs2015", "EmsApi.Client.Vs2015.csproj", "{60E9EF36-B92A-4486-A1C1-BD90C44188C4}"
-	ProjectSection(ProjectDependencies) = postProject
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D} = {86F1DF15-9BC1-4828-80DB-878DB1DB037D}
-	EndProjectSection
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{54654233-4D80-477A-8FA7-CD7AA299E402}"
-	ProjectSection(SolutionItems) = preProject
-		..\.editorconfig = ..\.editorconfig
-		..\..\appveyor.yml = ..\..\appveyor.yml
-		..\build.ps1 = ..\build.ps1
-		EmsApi.Client.nuspec = EmsApi.Client.nuspec
-		..\EmsApi.Dto\nswag.json = ..\EmsApi.Dto\nswag.json
-		..\README.md = ..\README.md
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Client.Tests", "..\Tests\EmsApi.Client.Tests.csproj", "{76BAA6AD-9B05-4581-A907-AC6FA2032512}"
-	ProjectSection(ProjectDependencies) = postProject
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4} = {60E9EF36-B92A-4486-A1C1-BD90C44188C4}
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Dto", "..\EmsApi.Dto\EmsApi.Dto.csproj", "{86F1DF15-9BC1-4828-80DB-878DB1DB037D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Release|Any CPU.Build.0 = Release|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
