--- csharp/EmsApi.Client/EmsApi.Client.sln
@@ -1,70 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26206.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Client", "EmsApi.Client.csproj", "{F924F980-8D7A-4616-8EEB-D5336C14BBED}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D7B2D7F6-0AA5-4260-A913-8FC87769EE93}"
-	ProjectSection(SolutionItems) = preProject
-		..\README.md = ..\README.md
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Client.Tests", "..\Tests\EmsApi.Client.Tests.csproj", "{76BAA6AD-9B05-4581-A907-AC6FA2032512}"
-	ProjectSection(ProjectDependencies) = postProject
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED} = {F924F980-8D7A-4616-8EEB-D5336C14BBED}
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Dto", "..\EmsApi.Dto\EmsApi.Dto.csproj", "{86F1DF15-9BC1-4828-80DB-878DB1DB037D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Debug|x64.ActiveCfg = Debug|x64
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Debug|x64.Build.0 = Debug|x64
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Debug|x86.ActiveCfg = Debug|x86
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Debug|x86.Build.0 = Debug|x86
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Release|x64.ActiveCfg = Release|x64
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Release|x64.Build.0 = Release|x64
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Release|x86.ActiveCfg = Release|x86
-		{F924F980-8D7A-4616-8EEB-D5336C14BBED}.Release|x86.Build.0 = Release|x86
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Debug|x64.Build.0 = Debug|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Debug|x86.Build.0 = Debug|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Release|Any CPU.Build.0 = Release|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Release|x64.ActiveCfg = Release|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Release|x64.Build.0 = Release|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Release|x86.ActiveCfg = Release|Any CPU
-		{76BAA6AD-9B05-4581-A907-AC6FA2032512}.Release|x86.Build.0 = Release|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|x64.Build.0 = Debug|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|x86.Build.0 = Debug|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|x64.ActiveCfg = Release|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|x64.Build.0 = Release|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|x86.ActiveCfg = Release|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
