+++ csharp/EmsApi.Dto/EmsApi.Dto.csproj
@@ -16,9 +16,8 @@
     <TargetFrameworkProfile>
     </TargetFrameworkProfile>
     <TargetFrameworkVersion>v5.0</TargetFrameworkVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(NuGetPackageRoot)' == ''">
-    <NuGetPackageRoot>$(UserProfile)\.nuget\packages\</NuGetPackageRoot>
+    <RestorePackages>true</RestorePackages>
+    <BuildPackage>true</BuildPackage>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -40,6 +39,9 @@
     <DocumentationFile>bin\Release\EmsApi.Dto.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
+    <None Include="EmsApi.Dto.nuspec">
+      <SubType>Designer</SubType>
+    </None>
     <None Include="project.json" />
     <!-- A reference to the entire .NET Framework is automatically included -->
   </ItemGroup>
@@ -48,21 +50,15 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
-  <Import Project="$(NuGetPackageRoot)\NSwag.MSBuild\9.5.0\build\NSwag.MSBuild.props" Condition="'$(NSwagExe)' == ''" />
+  <Import Project="$(SolutionDir)EmsApi.targets" />
   <Target Name="BeforeBuild">
     <Exec Command="$(NSwagExe) swagger2csclient /input:$(ProjectDir)ems-api.json /namespace:EmsApi.Dto.V2 /output:$(ProjectDir)Generated/EmsApi.Dto.V2.cs /generateClientClasses:false /generateDtoTypes:true /generateExceptionClasses:false /generateOptionalParameters:false /generateDataAnnotations:false /generateDefaultValues:false /classStyle:Poco" />
-    <!-- <Exec Command="$(NSwagExe) swagger2tsclient /input:$(ProjectDir)ems-api.json /namespace:ems.api.v2 /moduleName:ems.api.v2 /output:$(ProjectDir)Generated/ems-api-v2.d.ts /generateClientClasses:false /generateDtoTypes:true /generateDefaultValues:false /typeStyle:Interface" /> -->
   </Target>
   <PropertyGroup>
-    <PostBuildEvent>echo Copy to EmsApi.Dto\bin for nuget package
-copy "$(TargetPath)" ..\
-copy "$(TargetDir)EmsApi.Dto.xml" ..\
-
-echo Copy generated output to csharp\bin
-mkdir "$(ProjectDir)..\bin\"
-copy "$(TargetPath)" "$(ProjectDir)..\bin\"
-copy "$(ProjectDir)Generated\EmsApi.Dto.V2.cs" "$(ProjectDir)..\bin\"
-echo copy "$(ProjectDir)Generated\ems-api-v2.d.ts" "$(ProjectDir)..\bin\"</PostBuildEvent>
+    <PostBuildEvent>echo Copy generated output to csharp\bin
+mkdir "$(SolutionBin)"
+copy "$(ProjectDir)Generated\EmsApi.Dto.V2.cs" "$(SolutionBin)"
+echo copy "$(ProjectDir)Generated\ems-api-v2.d.ts" "$(SolutionBin)"</PostBuildEvent>
   </PropertyGroup>
   <PropertyGroup>
     <PreBuildEvent>
