--- csharp/EmsApi.Dto/EmsApi.Dto.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Dto", "EmsApi.Dto.csproj", "{86F1DF15-9BC1-4828-80DB-878DB1DB037D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
