+++ csharp/EmsApi.sln
@@ -1,0 +1,44 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Client", "EmsApi.Client\EmsApi.Client.csproj", "{60E9EF36-B92A-4486-A1C1-BD90C44188C4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Dto", "EmsApi.Dto\EmsApi.Dto.csproj", "{86F1DF15-9BC1-4828-80DB-878DB1DB037D}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{87FF617C-FF36-4CB6-A61D-7C90C203A241}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+		.gitignore = .gitignore
+		..\appveyor.yml = ..\appveyor.yml
+		build.ps1 = build.ps1
+		EmsApi.targets = EmsApi.targets
+		README.md = README.md
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Tests", "EmsApi.Tests\EmsApi.Tests.csproj", "{6695E626-5B8B-4A41-BF5D-E1BB8D46F162}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{86F1DF15-9BC1-4828-80DB-878DB1DB037D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6695E626-5B8B-4A41-BF5D-E1BB8D46F162}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6695E626-5B8B-4A41-BF5D-E1BB8D46F162}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6695E626-5B8B-4A41-BF5D-E1BB8D46F162}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6695E626-5B8B-4A41-BF5D-E1BB8D46F162}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
