+++ csharp/Examples/EmsApi.Example.DotnetFrameworkConsole/EmsApi.Example.DotnetFrameworkConsole.csproj
@@ -36,7 +36,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="EmsApi.Client">
-      <HintPath>..\..\EmsApi.Client\bin\EmsApi.Client.dll</HintPath>
+      <HintPath>..\..\bin\EmsApi.Client.dll</HintPath>
     </Reference>
     <Reference Include="Microsoft.AspNetCore.WebUtilities, Version=1.0.0.0, Culture=neutral, PublicKeyToken=adb9793829ddae60, processorArchitecture=MSIL">
       <HintPath>..\packages\Microsoft.AspNetCore.WebUtilities.1.0.0\lib\net451\Microsoft.AspNetCore.WebUtilities.dll</HintPath>
