--- csharp/Examples/EmsApi.Example.Vs2015.sln
@@ -1,40 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Example.DotnetFrameworkConsole", "EmsApi.Example.DotnetFrameworkConsole\EmsApi.Example.DotnetFrameworkConsole.csproj", "{4E34536D-89FB-4B1D-B47D-95DAE73EF66B}"
-	ProjectSection(ProjectDependencies) = postProject
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4} = {60E9EF36-B92A-4486-A1C1-BD90C44188C4}
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Example.WPF", "EmsApi.Example.WPF\EmsApi.Example.WPF.csproj", "{E601392F-C03F-4381-B4A5-4D62A080919E}"
-	ProjectSection(ProjectDependencies) = postProject
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4} = {60E9EF36-B92A-4486-A1C1-BD90C44188C4}
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmsApi.Client.Vs2015", "..\EmsApi.Client\EmsApi.Client.Vs2015.csproj", "{60E9EF36-B92A-4486-A1C1-BD90C44188C4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4E34536D-89FB-4B1D-B47D-95DAE73EF66B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4E34536D-89FB-4B1D-B47D-95DAE73EF66B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4E34536D-89FB-4B1D-B47D-95DAE73EF66B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4E34536D-89FB-4B1D-B47D-95DAE73EF66B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E601392F-C03F-4381-B4A5-4D62A080919E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E601392F-C03F-4381-B4A5-4D62A080919E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E601392F-C03F-4381-B4A5-4D62A080919E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E601392F-C03F-4381-B4A5-4D62A080919E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{60E9EF36-B92A-4486-A1C1-BD90C44188C4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
